+++ src/lib/game/items/tool.ts
@@ -1,4 +1,4 @@
-type materialType = 'wood' | 'stone' | 'coal' | 'iron' | 'redstone' | 'gold' | 'diamond'
+export type materialType = 'wood' | 'stone' | 'coal' | 'iron' | 'redstone' | 'gold' | 'diamond'
 
 export interface Item {
     emote: string;
@@ -8,5 +8,7 @@ export interface Item {
 export interface Tool extends Item {
     durability: number;
     enchants: string[];
+    xp: [number, number, number];
     materials: Record<materialType, number>;
+    repair: Record<materialType | string, number>;
 }